Set kallisto index length dynamically
Use the length of the two index reads to determine if we should run
SwabSeq or SwabSeq10 for kallisto bus.

Also build whitelist directly from SampleSheet.csv instead of joining in
barcode map information (not required for kallisto). Wondering if
SampleSheet is the best way to go through this - maybe user could touch
a file that specifies one of our set whitelists?